+++ cypress/cypress/component/Typeahead/StaticTypeaheadInput.cy.tsx
@@ -1,10 +1,19 @@
+/* eslint-disable max-lines */
 import { Form, StaticTypeaheadInput } from "react-hook-form-components";
 import { faker, Sex } from "@faker-js/faker";
 import * as yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { generateOptions } from "../../helpers/typeahead";
-import { useEffect, useRef } from "react";
-import TypeheadRef from "react-bootstrap-typeahead/types/core/Typeahead";
+import { useState } from "react";
+
+const selectOption = (name: string, text: string) => {
+  cy.get(`#${name}`).clear().click().type(text);
+  cy.get('li[role="option"]').contains(text).click();
+};
+
+const waitForChip = (text: string) => {
+  cy.get(`span.MuiChip-label:contains(${text})`).should("be.visible");
+};
 
 it("works with single simple options", () => {
   const { simpleOptions } = generateOptions();
@@ -14,173 +23,138 @@ it("works with single simple options", () => {
   const [defaultSelectedOption, changedOption] = randomOptions;
 
   cy.mount(
-    <Form
-      onSubmit={cy.spy().as("onSubmitSpy")}
-      defaultValues={{
-        [name]: defaultSelectedOption,
-      }}
-    >
-      <StaticTypeaheadInput name={name} label={name} options={simpleOptions} defaultSelected={[defaultSelectedOption]} />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form
+        onSubmit={cy.spy().as("onSubmitSpy")}
+        defaultValues={{
+          [name]: defaultSelectedOption,
+        }}
+      >
+        <StaticTypeaheadInput name={name} label={name} options={simpleOptions} />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
   cy.get(`#${name}`).should("have.value", defaultSelectedOption);
   cy.get("input[type=submit]").click({ force: true });
   cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: defaultSelectedOption });
 
-  cy.get(`#${name}`).clear();
-  cy.get(`#${name}`).click();
-  cy.get(`a[aria-label='${changedOption}']`).click();
+  selectOption(name, changedOption);
   cy.get("input[type=submit]").click({ force: true });
   cy.get("@onSubmitSpy").should("be.calledWith", { [name]: changedOption });
 });
 
-it("select automatically a single simple options - single", () => {
+it("select automatically a single simple options - single (autoSelect/autoHighlight)", () => {
   const { simpleOptions } = generateOptions();
   const name = faker.random.alpha(10);
   const randomOptions = faker.helpers.arrayElements(simpleOptions, 2);
   const [firstOption] = randomOptions;
 
   cy.mount(
-    <Form onSubmit={cy.spy().as("onSubmitSpy")}>
-      <StaticTypeaheadInput name={name} label={name} options={simpleOptions} />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form onSubmit={cy.spy().as("onSubmitSpy")}>
+        <StaticTypeaheadInput name={name} label={name} options={simpleOptions} autoSelect />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
   cy.get(`#${name}`).type(firstOption);
-  cy.get(`#${name}`).blur({ force: true });
-  // eslint-disable-next-line cypress/no-unnecessary-waiting
-  cy.wait(100);
-  cy.get("div[class=invalid-feedback]").should("not.be.visible");
+  cy.get(`#${name}`).blur();
+  cy.get("ul.MuiAutocomplete-listbox").should("not.exist");
   cy.get("input[type=submit]").click({ force: true });
   cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: firstOption });
 });
 
-it("display an error if more than one options are found and not selected - single", () => {
+it("select automatically a single simple options - multiple (when autoSelect/autoHighlight)", () => {
   const { simpleOptions } = generateOptions();
   const name = faker.random.alpha(10);
   const randomOptions = faker.helpers.arrayElements(simpleOptions, 2);
-  const [firstOption] = randomOptions;
-  const errorMessage = faker.random.words(3);
-
-  const additionalOption = firstOption + "xyz";
 
   cy.mount(
-    <Form onSubmit={cy.spy().as("onSubmitSpy")}>
-      <StaticTypeaheadInput name={name} label={name} options={[...simpleOptions, additionalOption]} invalidErrorMessage={errorMessage} />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form onSubmit={cy.spy().as("onSubmitSpy")}>
+        <StaticTypeaheadInput name={name} label={name} options={simpleOptions} multiple autoSelect />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
-  cy.get(`#${name}`).type(firstOption);
-  cy.get(`#${name}`).blur({ force: true });
-  // eslint-disable-next-line cypress/no-unnecessary-waiting
-  cy.wait(100);
-  cy.get("div[class=invalid-feedback]").should("be.visible").should("have.text", errorMessage);
+  cy.get(`#${name}`).type(randomOptions[0]);
+  cy.get(`#${name}`).blur();
+  waitForChip(randomOptions[0]);
+
+  cy.get(`#${name}`).type(randomOptions[1]);
+  cy.get(`#${name}`).blur();
+  waitForChip(randomOptions[1]);
+
   cy.get("input[type=submit]").click({ force: true });
-  cy.get("@onSubmitSpy").should("not.have.been.called");
-  cy.get("div[class=invalid-feedback]").should("be.visible").should("have.text", errorMessage);
+  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: [randomOptions[0], randomOptions[1]] });
 });
 
-it("select automatically a single simple options - multiple", () => {
-  const { simpleOptions } = generateOptions();
+it("select automatically single option when multiple options are available - single (autoSelect/autoHighlight)", () => {
+  const prefix = faker.random.alpha(3);
+  const options = [
+    { label: prefix + faker.random.alpha(7), value: faker.datatype.uuid() },
+    { label: prefix + faker.random.alpha(7), value: faker.datatype.uuid() },
+  ];
   const name = faker.random.alpha(10);
-  const randomOptions = faker.helpers.arrayElements(simpleOptions, 2);
-  const [firstOption, secondOption] = randomOptions;
 
   cy.mount(
-    <Form onSubmit={cy.spy().as("onSubmitSpy")}>
-      <StaticTypeaheadInput name={name} label={name} options={simpleOptions} multiple />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form onSubmit={cy.spy().as("onSubmitSpy")}>
+        <StaticTypeaheadInput name={name} label={name} options={options} autoSelect />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
-  cy.get(`#${name}`).type(firstOption);
-  cy.get(`#${name}`).blur({ force: true });
-  // eslint-disable-next-line cypress/no-unnecessary-waiting
-  cy.wait(100);
-  cy.get(`#${name}`).type(secondOption);
-  cy.get(`#${name}`).blur({ force: true });
-  // eslint-disable-next-line cypress/no-unnecessary-waiting
-  cy.wait(100);
-  cy.get("div[class=invalid-feedback]").should("not.be.visible");
+  cy.get(`#${name}`).type(prefix);
+  cy.get(`#${name}`).type("{downarrow}");
+  cy.get(`#${name}`).blur();
+  cy.get("ul.MuiAutocomplete-listbox").should("not.exist");
   cy.get("input[type=submit]").click({ force: true });
-  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: [firstOption, secondOption] });
-});
-
-it("display an error if more than one options are found and not selected - multiple", () => {
-  const { simpleOptions } = generateOptions();
-  const name = faker.random.alpha(10);
-  const randomOptions = faker.helpers.arrayElements(simpleOptions, 2);
-  const [firstOption, secondOption] = randomOptions;
-  const errorMessage = faker.random.words(3);
-
-  const additionalOption = secondOption + "xyz";
-
-  cy.mount(
-    <Form onSubmit={cy.spy().as("onSubmitSpy")}>
-      <StaticTypeaheadInput
-        name={name}
-        label={name}
-        options={[...simpleOptions, additionalOption]}
-        multiple
-        invalidErrorMessage={errorMessage}
-      />
-      <input type="submit" />
-    </Form>,
-  );
-
-  cy.get(`#${name}`).type(firstOption);
-  cy.get(`#${name}`).blur({ force: true });
-  // eslint-disable-next-line cypress/no-unnecessary-waiting
-  cy.wait(100);
-  cy.get(`#${name}`).type(secondOption);
-  cy.get(`#${name}`).blur({ force: true });
-  // eslint-disable-next-line cypress/no-unnecessary-waiting
-  cy.wait(100);
-  cy.get("div[class=invalid-feedback]").should("exist");
-  cy.get("input[type=submit]").click({ force: true });
-  cy.get("@onSubmitSpy").should("not.have.been.called");
-  cy.get("div[class=invalid-feedback]").should("be.visible").should("have.text", errorMessage);
+  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: options[1].value });
 });
 
 it("works with multiple simple options", () => {
-  const { simpleOptions } = generateOptions();
+  const options = generateOptions();
   const name = faker.random.alpha(10);
-
-  const randomOptions = faker.helpers.arrayElements(simpleOptions, faker.datatype.number({ min: 2, max: 5 }));
+  const randomOptions = faker.helpers.arrayElements(options.objectOptions, faker.datatype.number({ min: 2, max: 5 }));
   const half = Math.ceil(randomOptions.length / 2);
 
   const defaultSelectedOptions = randomOptions.slice(0, half);
   const changedOptions = randomOptions.slice(half);
 
   cy.mount(
-    <Form
-      onSubmit={cy.spy().as("onSubmitSpy")}
-      defaultValues={{
-        [name]: defaultSelectedOptions,
-      }}
-    >
-      <StaticTypeaheadInput multiple name={name} label={name} options={simpleOptions} defaultSelected={defaultSelectedOptions} />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form
+        onSubmit={cy.spy().as("onSubmitSpy")}
+        defaultValues={{
+          [name]: defaultSelectedOptions.map((o) => o.value),
+        }}
+      >
+        <StaticTypeaheadInput multiple name={name} label={name} options={options.objectOptions} />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
   cy.get("input[type=submit]").click({ force: true });
-  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: defaultSelectedOptions });
+  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: defaultSelectedOptions.map((o) => o.value) });
 
-  cy.get(`#${name}`).click();
-  cy.get(`#${name}`).type("{backspace}".repeat(20));
+  cy.get(`#${name}`).click().type("{backspace}".repeat(20));
 
-  for (const option of changedOptions) {
-    cy.get(`#${name}`).click();
-    cy.get(`a[aria-label='${option}']`).click();
+  for (const changedOption of changedOptions) {
+    selectOption(name, changedOption.label);
+    waitForChip(changedOption.label);
   }
 
+  cy.get(`#${name}`).blur();
+  cy.get("ul.MuiAutocomplete-listbox").should("not.exist");
   cy.get("input[type=submit]").click({ force: true });
-  cy.get("@onSubmitSpy").should("be.calledWith", { [name]: changedOptions });
+  cy.get("@onSubmitSpy").should("be.calledWith", { [name]: changedOptions.map((o) => o.value) });
 });
 
 it("works with single object options", () => {
@@ -189,24 +163,25 @@ it("works with single object options", () => {
   const [defaultSelectedOption, changedOption] = faker.helpers.arrayElements(objectOptions, 2);
 
   cy.mount(
-    <Form
-      onSubmit={cy.spy().as("onSubmitSpy")}
-      defaultValues={{
-        [name]: defaultSelectedOption.value,
-      }}
-    >
-      <StaticTypeaheadInput name={name} label={name} options={objectOptions} defaultSelected={[defaultSelectedOption.label]} />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form
+        onSubmit={cy.spy().as("onSubmitSpy")}
+        defaultValues={{
+          [name]: defaultSelectedOption.value,
+        }}
+      >
+        <StaticTypeaheadInput name={name} label={name} options={objectOptions} />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
   cy.get(`#${name}`).should("have.value", defaultSelectedOption.label);
   cy.get("input[type=submit]").click({ force: true });
   cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: defaultSelectedOption.value });
 
-  cy.get(`#${name}`).clear();
-  cy.get(`#${name}`).click();
-  cy.get(`a[aria-label='${changedOption.label}']`).click();
+  cy.get(`#${name}`).clear().click();
+  selectOption(name, changedOption.label);
   cy.get("input[type=submit]").click({ force: true });
   cy.get("@onSubmitSpy").should("be.calledWith", { [name]: changedOption.value });
 });
@@ -221,15 +196,17 @@ it("works with multiple object options", () => {
   const changedOptions = randomOptions.slice(half);
 
   cy.mount(
-    <Form
-      onSubmit={cy.spy().as("onSubmitSpy")}
-      defaultValues={{
-        [name]: defaultSelectedOptions.map((o) => o.value),
-      }}
-    >
-      <StaticTypeaheadInput multiple defaultSelected={defaultSelectedOptions} name={name} label={name} options={options.objectOptions} />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form
+        onSubmit={cy.spy().as("onSubmitSpy")}
+        defaultValues={{
+          [name]: defaultSelectedOptions.map((o) => o.value),
+        }}
+      >
+        <StaticTypeaheadInput multiple name={name} label={name} options={options.objectOptions} />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
   cy.get("input[type=submit]").click({ force: true });
@@ -239,11 +216,12 @@ it("works with multiple object options", () => {
   cy.get(`#${name}`).type("{backspace}".repeat(20));
 
   for (const changedOption of changedOptions) {
-    cy.get(`#${name}`).click();
-    cy.get(`#${name}`).type(changedOption.label);
-    cy.get(`a[aria-label='${changedOption.label}']`).click();
+    selectOption(name, changedOption.label);
+    waitForChip(changedOption.label);
   }
 
+  cy.get(`#${name}`).blur();
+  cy.get("ul.MuiAutocomplete-listbox").should("not.exist");
   cy.get("input[type=submit]").click({ force: true });
   cy.get("@onSubmitSpy").should("be.calledWith", { [name]: changedOptions.map((o) => o.value) });
 });
@@ -259,17 +237,18 @@ it("Validation works", () => {
   });
 
   cy.mount(
-    <Form onSubmit={cy.spy().as("onSubmitSpy")} resolver={yupResolver(schema)}>
-      <StaticTypeaheadInput name={name} label={name} options={simpleOptions} />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form onSubmit={cy.spy().as("onSubmitSpy")} resolver={yupResolver(schema)}>
+        <StaticTypeaheadInput name={name} label={name} options={simpleOptions} />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
   cy.get("input[type=submit]").click({ force: true });
   cy.contains(errorMessage).should("exist");
 
-  cy.get(`#${name}`).click();
-  cy.contains("a", randomOption).click();
+  selectOption(name, randomOption);
   cy.get("input[type=submit]").click({ force: true });
   cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: randomOption });
 });
@@ -280,37 +259,59 @@ it("works with the correct value onChange", () => {
   const randomOption = faker.helpers.arrayElement(simpleOptions);
 
   cy.mount(
-    <Form
-      onSubmit={() => {
-        // Nothing to do
-      }}
-    >
-      <StaticTypeaheadInput name={name} label={name} options={simpleOptions} onChange={cy.spy().as("OnChangeSpy")} />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form
+        onSubmit={() => {
+          // Nothing to do
+        }}
+      >
+        <StaticTypeaheadInput name={name} label={name} options={simpleOptions} onChange={cy.spy().as("OnChangeSpy")} />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
-  cy.get(`#${name}`).clear();
-  cy.get(`#${name}`).click();
-  cy.get(`a[aria-label='${randomOption}']`).click();
+  selectOption(name, randomOption);
   cy.get("@OnChangeSpy").should("have.been.calledWith", randomOption);
 });
 
-it("is disabled", () => {
+it("it is disabled", () => {
   const name = faker.random.word();
-  const { simpleOptions } = generateOptions(100);
+  const { simpleOptions } = generateOptions();
 
   cy.mount(
-    <Form
-      onSubmit={() => {
-        // Do nothing
-      }}
-    >
-      <StaticTypeaheadInput name={name} label={name} options={simpleOptions} disabled />
-    </Form>,
+    <div className="p-4">
+      <Form
+        onSubmit={() => {
+          // Do nothing
+        }}
+      >
+        <StaticTypeaheadInput name={name} label={name} options={simpleOptions} disabled />
+      </Form>
+    </div>,
   );
 
-  cy.get("input.rbt-input-main").should("be.disabled");
+  cy.get("input").should("be.disabled");
+});
+
+it("it is readonly", () => {
+  const name = faker.random.word();
+  const { simpleOptions } = generateOptions();
+
+  cy.mount(
+    <div className="p-4">
+      <Form
+        onSubmit={() => {
+          // Do nothing
+        }}
+      >
+        <StaticTypeaheadInput name={name} label={name} options={simpleOptions} readOnly />
+      </Form>
+    </div>,
+  );
+
+  cy.get("input").should("have.attr", "readonly");
+  cy.get(`#${name}`).should("have.value", "");
 });
 
 it("auto mark on focus", () => {
@@ -319,79 +320,76 @@ it("auto mark on focus", () => {
   const randomOption = faker.helpers.arrayElement(simpleOptions);
 
   cy.mount(
-    <Form
-      onSubmit={() => {
-        // Do nothing
-      }}
-    >
-      <StaticTypeaheadInput defaultSelected={[randomOption]} name={name} label={name} options={simpleOptions} markAllOnFocus />
-    </Form>,
+    <div className="p-4">
+      <Form
+        onSubmit={() => {
+          // Do nothing
+        }}
+        defaultValues={{
+          [name]: randomOption,
+        }}
+      >
+        <StaticTypeaheadInput name={name} label={name} options={simpleOptions} markAllOnFocus />
+      </Form>
+    </div>,
   );
 
   cy.contains("label", name).click();
   cy.get(`input[id=${name}]`).getSelectedText().should("eq", randomOption);
 });
 
-it("disabled options", () => {
+it("try to select a disabled option", () => {
   const { disabledOptions } = generateOptions();
   const name = faker.random.alpha(10);
-
-  const randomOptions = faker.helpers.arrayElements(disabledOptions, faker.datatype.number({ min: 2, max: 5 }));
-  const half = Math.ceil(randomOptions.length / 2);
-
-  const defaultSelectedOptions = randomOptions.slice(0, half);
-  const changedOptions = randomOptions.slice(half);
+  const randomOptions = faker.helpers.arrayElements(disabledOptions, 1);
+  const [randomOption] = randomOptions;
 
   cy.mount(
-    <Form
-      onSubmit={cy.spy().as("onSubmitSpy")}
-      defaultValues={{
-        [name]: defaultSelectedOptions,
-      }}
-    >
-      <StaticTypeaheadInput multiple name={name} label={name} options={disabledOptions} defaultSelected={defaultSelectedOptions} />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form
+        onSubmit={() => {
+          // Nothing to do
+        }}
+      >
+        <StaticTypeaheadInput multiple name={name} label={name} options={randomOptions} />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
-  cy.get("input[type=submit]").click({ force: true });
-  cy.get("@onSubmitSpy").should("be.calledOnceWith", { [name]: defaultSelectedOptions });
-
-  cy.get(`#${name}`).click();
-  cy.get(`#${name}`).type("{backspace}".repeat(20));
-
-  for (const option of changedOptions) {
-    cy.get(`#${name}`).click();
-    cy.get(`a[aria-label='${option.label}']`).should("have.class", "disabled");
-  }
+  cy.get(`#${name}`).clear().click().type(randomOption.label);
+  cy.get('li[role="option"]').should("have.attr", "aria-disabled", "true");
 });
 
-it("empty label", () => {
+it("test empty label", () => {
   const { simpleOptions } = generateOptions();
   const name = faker.random.alpha(10);
   const emptyLabel = faker.random.words(5);
 
   cy.mount(
-    <Form
-      onSubmit={cy.spy().as("onSubmitSpy")}
-      defaultValues={{
-        [name]: simpleOptions,
-      }}
-    >
-      <StaticTypeaheadInput
-        multiple
-        name={name}
-        label={name}
-        options={simpleOptions}
-        defaultSelected={simpleOptions}
-        emptyLabel={emptyLabel}
-      />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form
+        onSubmit={cy.spy().as("onSubmitSpy")}
+        defaultValues={{
+          [name]: simpleOptions,
+        }}
+      >
+        <StaticTypeaheadInput
+          multiple
+          name={name}
+          label={name}
+          options={simpleOptions}
+          autocompleteProps={{
+            noOptionsText: emptyLabel,
+          }}
+        />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
-  cy.get(`#${name}`).click();
-  cy.get(".dropdown-menu > .dropdown-item").should("have.text", emptyLabel);
+  cy.get(`#${name}`).clear().click().type(name);
+  cy.get("div.MuiAutocomplete-noOptions").should("have.text", emptyLabel);
 });
 
 it("placeholder", () => {
@@ -400,71 +398,80 @@ it("placeholder", () => {
   const placeholder = faker.random.words(5);
 
   cy.mount(
-    <Form
-      onSubmit={cy.spy().as("onSubmitSpy")}
-      defaultValues={{
-        [name]: simpleOptions,
-      }}
-    >
-      <StaticTypeaheadInput multiple name={name} label={name} options={simpleOptions} placeholder={placeholder} />
-      <input type="submit" />
-    </Form>,
-  );
-  cy.get(`#${name}`).should("have.attr", "placeholder", placeholder);
-});
-
-it("use input-ref", () => {
-  const { simpleOptions } = generateOptions();
-  const name = faker.random.alpha(10);
-  const placeholder = faker.random.words(5);
-
-  const TestForm = () => {
-    const ref = useRef<TypeheadRef | null>(null);
-
-    useEffect(() => {
-      if (ref.current) {
-        ref.current.toggleMenu();
-      }
-    }, [ref]);
-
-    return (
+    <div className="p-4">
       <Form
         onSubmit={cy.spy().as("onSubmitSpy")}
         defaultValues={{
           [name]: simpleOptions,
         }}
       >
-        <StaticTypeaheadInput inputRef={ref} name={name} label={name} options={simpleOptions} placeholder={placeholder} />
+        <StaticTypeaheadInput multiple name={name} label={name} options={simpleOptions} placeholder={placeholder} />
+        <input type="submit" className="mt-4" />
       </Form>
+    </div>,
+  );
+
+  cy.get(`#${name}`).should("have.attr", "placeholder", placeholder);
+});
+
+it("test on input change", () => {
+  const { simpleOptions } = generateOptions();
+  const name = faker.random.alpha(10);
+  const text = faker.random.words(5);
+
+  const TestForm = () => {
+    const [disabled, setDisabled] = useState<boolean>(false);
+    return (
+      <div className="p-4">
+        <Form
+          onSubmit={cy.spy().as("onSubmitSpy")}
+          defaultValues={{
+            [name]: simpleOptions,
+          }}
+        >
+          <StaticTypeaheadInput
+            name={name}
+            label={name}
+            options={simpleOptions}
+            onInputChange={(text: string) => setDisabled(text.length === 0)}
+          />
+          <input type="submit" className="mt-4" disabled={disabled} />
+        </Form>
+      </div>
     );
   };
 
   cy.mount(<TestForm />);
-  cy.get(".rbt-menu.dropdown-menu.show").should("be.visible");
+  cy.get(`#${name}`).clear().click().type(text);
+  cy.get('input[type="submit"]').should("be.enabled");
+  cy.get(`#${name}`).clear();
+  cy.get('input[type="submit"]').should("be.disabled");
 });
 
-it("grouping options", () => {
+it("test grouping options", () => {
   const COUNT = 10;
   const { groupedOptions } = generateOptions(COUNT);
   const name = faker.random.alpha(10);
 
   cy.mount(
-    <Form
-      onSubmit={() => {
-        // Nothing to do
-      }}
-    >
-      <StaticTypeaheadInput useGroupBy name={name} label={name} options={groupedOptions} />
-      <input type="submit" />
-    </Form>,
+    <div className="p-4">
+      <Form
+        onSubmit={() => {
+          // Nothing to do
+        }}
+      >
+        <StaticTypeaheadInput name={name} label={name} useGroupBy options={groupedOptions} />
+        <input type="submit" className="mt-4" />
+      </Form>
+    </div>,
   );
 
-  cy.get(`#${name}`).click();
-  cy.get(".dropdown-menu.show")
-    .find("a")
-    .should("have.length", COUNT + 1);
-  cy.get(".dropdown-header").first().should("be.visible").and("have.text", Sex.Male);
-  cy.get(".dropdown-header").eq(1).should("be.visible").and("have.text", Sex.Female);
-  cy.contains("a", groupedOptions[0].label).should("have.class", "disabled");
-  cy.contains("a", groupedOptions[COUNT].label).should("exist");
+  cy.get(`#${name}`).type(groupedOptions[0].label);
+  cy.get("div.MuiAutocomplete-groupLabel").first().should("be.visible").and("have.text", Sex.Male);
+  cy.get(`#${name}`)
+    .clear()
+    .type(groupedOptions[COUNT / 2].label);
+  cy.get("div.MuiAutocomplete-groupLabel").first().should("be.visible").and("have.text", Sex.Female);
+  cy.get(`#${name}`).clear().type(groupedOptions[COUNT].label);
+  cy.get('li[role="option"]').contains(groupedOptions[COUNT].label).should("exist");
 });