+++ cypress/cypress/support/component.ts
@@ -14,7 +14,6 @@
 // ***********************************************************
 
 import "bootstrap/dist/css/bootstrap.min.css";
-import "react-bootstrap-typeahead/css/Typeahead.css";
 import "react-datepicker/dist/react-datepicker.css";
 import "react-hook-form-components/styles.css";